lit_cli/requirements.txt
build==1.2.2.post1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
docutils==0.21.2
halo==0.0.31
id==1.5.0
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
keyring==25.6.0
log-symbols==0.0.14
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.7.0
nh3==0.2.21
packaging==25.0
pycparser==2.22
Pygments==2.19.1
pyproject_hooks==1.2.0
PyYAML==6.0.2
readme_renderer==44.0
requests==2.32.4
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
SecretStorage==3.3.3
six==1.17.0
spinners==0.0.24
termcolor==3.1.0
toml==0.10.2
twine==6.1.0
urllib3==2.4.0
